requirements.txt
asgiref==3.8.1
Django==4.2.10
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-mssql-backend==2.8.1
gunicorn==23.0.0
mssql-django==1.5
packaging==25.0
pyodbc==5.2.0
pytz==2025.2
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.2